+++ parser_library/test/expressions/character_expression_test.cpp
@@ -12,6 +12,8 @@
  *   Broadcom, Inc. - initial API and implementation
  */
 
+#include <tuple>
+
 #include "gtest/gtest.h"
 
 #include "../common_testing.h"
@@ -343,25 +345,32 @@ TEST(character_expression, bounds)
 {
     const std::string input_template = R"(
 &C  SETC  'A'      
-&C  SETC  '&C'%
+&C  SETC  #'&C'%
 )";
-    for (const auto& [replacement, err] : std::initializer_list<std::pair<std::string_view, std::string_view>> {
-             { "(0,*)", "CE008" },
-             { "(1,*)", "" },
-             { "(2,*)", "" },
-             { "(2,1)", "CE009" },
-             { "(3,*)", "CE009" },
-         })
+    for (const auto& [prefix, suffix, err] :
+        std::vector<std::tuple<std::string_view, std::string_view, std::vector<std::string_view>>> {
+            { "", "(0,*)", { "CE008" } },
+            { "", "(1,*)", {} },
+            { "", "(2,*)", {} },
+            { "", "(2,1)", { "CE009" } },
+            { "", "(3,*)", { "CE009" } },
+            { "(-1)", "", { "CE010" } },
+            { "(-1)", "(1,*)", { "CE010" } },
+            { "(-1)", "(1,-1)", { "CE008", "CE010" } },
+            { "(-1)", "(-1,*)", { "CE008", "CE010" } },
+            { "(-1)", "(-1,-1)", { "CE008", "CE010" } },
+            { "(0)", "(-1,-1)", { "CE008" } },
+            { "(4064)", "(1,1)", {} },
+            { "(4065)", "(1,1)", { "CE011" } },
+        })
     {
         std::string input = input_template;
-        input.replace(input.find('%'), 1, replacement);
+        input.replace(input.find('#'), 1, prefix);
+        input.replace(input.find('%'), 1, suffix);
 
         analyzer a(input);
         a.analyze();
 
-        if (err.empty())
-            EXPECT_TRUE(a.diags().empty());
-        else
-            EXPECT_TRUE(matches_message_codes(a.diags(), { std::string(err) }));
+        EXPECT_TRUE(matches_message_codes(a.diags(), err));
     }
 }